# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LoreSafe API",
    "version": "1.0.0",
    "description": "Versioned contract for LoreSafe's spoiler-safe REST and SSE API. Governance and compatibility rules live in context/api-governance.md."
  },
  "servers": [
    {
      "url": "https://api.loresafe.org"
    }
  ],
  "security": [
    {
      "cookieSession": []
    }
  ],
  "paths": {
...
    "/api/auth/password/reset": {
      "post": {
        "operationId": "resetPassword",
        "summary": "reset password",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "x-response-dto": "CompletedResponseDto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-idempotent": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "loresafe_session",
        "description": "Short-lived JWT stored only in a Secure, HttpOnly cookie."
      },
      "operationsBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dedicated operations token required only for metrics."
      }
    },
    "schemas": {
...
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 40,
            "maxLength": 128
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "token",
          "password"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "description": "Server request correlation ID. Client x-request-id values are accepted only as 1-64 character trace-compatible tokens containing letters, digits, dot, underscore, colon, or hyphen; invalid values are replaced with a server UUID. Correlation IDs are not uniqueness or authorization keys."
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "Authenticated but not authorized.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Conflict": {
        "description": "Request conflicts with current resource state.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded; retry after the advertised reset.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "RateLimit": {
            "description": "Structured limit, remaining quota, and reset parameters.",
            "schema": {
              "type": "string"
            }
          },
          "Retry-After": {
            "description": "Seconds until a retry is permitted when present.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A required dependency is temporarily unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    }
  }
}